Conflicts:
	components/tools/OmeroWeb/omeroweb/webclient/templates/webclient/data/includes/center_plugin.table.js.html

+++ components/tools/OmeroPy/src/omero/gateway/__init__.py
@@ -25,7 +25,8 @@ import ConfigParser
 import omero
 import omero.clients
 from omero.util.decorators import timeit, TimeIt, setsessiongroup
-from omero.cmd import Chgrp
+from omero.cmd import Chgrp, DoAll
+from omero.api import Save
 from omero.callbacks import CmdCallbackI
 from omero.gateway.utils import ServiceOptsDict, GatewayConfig
 import omero.scripts as scripts
@@ -742,12 +743,15 @@ class BlitzObjectWrapper (object):
 
         # Using omero.cmd.Delete rather than deleteObjects since we need
         # spec/id pairs rather than spec+id_list as arguments
-        handle = self._conn.c.sf.submit(dcs, self._conn.SERVICE_OPTS)
-        callback = omero.callbacks.CmdCallbackI(self._conn.c, handle)
-        # Maximum wait time 5 seconds, will raise a LockTimeout if the
-        # delete has not finished by then.
-        callback.loop(10, 500)
-        self._obj.unloadAnnotationLinks()
+        if len(dcs):
+            doall = omero.cmd.DoAll()
+            doall.requests = dcs
+            handle = self._conn.c.sf.submit(doall, self._conn.SERVICE_OPTS)
+            try:
+                self._conn._waitOnCmd(handle)
+            finally:
+                handle.close()
+            self._obj.unloadAnnotationLinks()
 
     def removeAnnotations (self, ns):
         """
@@ -763,12 +767,13 @@ class BlitzObjectWrapper (object):
         for al in self._getAnnotationLinks(ns=ns):
             a = al.child
             ids.append(a.id.val)
-        handle = self._conn.deleteObjects('/Annotation', ids)
-        callback = omero.callbacks.CmdCallbackI(self._conn.c, handle)
-        # Maximum wait time 5 seconds, will raise a LockTimeout if the
-        # delete has not finished by then.
-        callback.loop(10, 500)
-        self._obj.unloadAnnotationLinks()        
+        if len(ids):
+            handle = self._conn.deleteObjects('/Annotation', ids)
+            try:
+                self._conn._waitOnCmd(handle)
+            finally:
+                handle.close()
+            self._obj.unloadAnnotationLinks()        
     
     # findAnnotations(self, ns=[])
     def getAnnotation (self, ns=None):
@@ -3152,16 +3157,16 @@ class _BlitzGateway (object):
         return prx
 
 
-    def chgrpObject(self, graph_spec, obj_id, group_id):
+    def chgrpObjects(self, graph_spec, obj_ids, group_id, container_id=None):
         """
-        Change the Group for a specified object using queue.
+        Change the Group for a specified objects using queue.
 
         @param graph_spec:      String to indicate the object type or graph
                                 specification. Examples include:
                                  * '/Image'
                                  * '/Project'   # will move contents too.
                                  * NB: Also supports 'Image' etc for convenience
-        @param obj_id:          ID for the object to move.
+        @param obj_ids:         IDs for the objects to move.
         @param group_id:        The group to move the data to.
         """
 
@@ -3169,9 +3174,31 @@ class _BlitzGateway (object):
             graph_spec = '/%s' % graph_spec
             logger.debug('chgrp Received object type, using "%s"' % graph_spec)
 
-        chgrp = omero.cmd.Chgrp(type=graph_spec, id=obj_id, options=None, grp=group_id)
+        # (link, child, parent)
+        parentLinkClasses = {"/Image": (omero.model.DatasetImageLinkI, omero.model.ImageI, omero.model.DatasetI),
+                        "/Dataset": (omero.model.ProjectDatasetLinkI, omero.model.DatasetI, omero.model.ProjectI),
+                        "/Plate": (omero.model.ScreenPlateLinkI, omero.model.PlateI, omero.model.ScreenI)}
+        da = DoAll()
+        requests = []
+        for obj_id in obj_ids:
+            obj_id = long(obj_id)
+            logger.debug('DoAll Chgrp: type: %s, id: %s, grp: %s' % (graph_spec, obj_id, group_id))
+            chgrp = omero.cmd.Chgrp(type=graph_spec, id=obj_id, options=None, grp=group_id)
+            requests.append(chgrp)
+            if container_id is not None and graph_spec in parentLinkClasses:
+                # get link class for graph_spec objects
+                link_klass = parentLinkClasses[graph_spec][0]
+                link = link_klass()
+                link.child = parentLinkClasses[graph_spec][1](obj_id, False)
+                link.parent = parentLinkClasses[graph_spec][2](container_id, False)
+                save = Save()
+                save.obj = link
+                requests.append(save)
 
-        prx = self.c.sf.submit(chgrp)
+        da.requests = requests
+        ctx = self.SERVICE_OPTS.copy()
+        ctx.setOmeroGroup(group_id)         # NB: For Save to work, we need to be in target group
+        prx = self.c.sf.submit(da, ctx)
         return prx
 
 
@@ -6033,6 +6060,13 @@ class _ImageWrapper (BlitzObjectWrapper):
             return (0, pmax-1)
 
     @assert_pixels
+    def requiresPixelsPyramid (self):
+        pixels_id = self._obj.getPrimaryPixels().getId().val
+        rp = self._conn.createRawPixelsStore()
+        rp.setPixelsId(pixels_id, True, self._conn.SERVICE_OPTS)
+        return rp.requiresPixelsPyramid()
+
+    @assert_pixels
     def getPrimaryPixels (self):
         """
         Loads pixels and returns object in a L{PixelsWrapper}
@@ -6068,16 +6102,19 @@ class _ImageWrapper (BlitzObjectWrapper):
         @type windows:      List of tuples. [(20, 300), (None, None), (50, 500)]. Must be tuples for all channels
         @param colors:      List of colors. ['F00', None, '00FF00'].  Must be item for each channel
         """
-
+        abs_channels = [abs(c) for c in channels]
+        idx = 0     # index of windows/colors args above
         for c in range(len(self.getChannels())):
             self._re.setActive(c, (c+1) in channels, self._conn.SERVICE_OPTS)
             if (c+1) in channels:
-                if windows is not None and windows[c][0] is not None and windows[c][1] is not None:
-                    self._re.setChannelWindow(c, *(windows[c] + [self._conn.SERVICE_OPTS]))
-                if colors is not None and colors[c]:
-                    rgba = splitHTMLColor(colors[c])
+                if windows is not None and windows[idx][0] is not None and windows[idx][1] is not None:
+                    self._re.setChannelWindow(c, *(windows[idx] + [self._conn.SERVICE_OPTS]))
+                if colors is not None and colors[idx]:
+                    rgba = splitHTMLColor(colors[idx])
                     if rgba:
                         self._re.setRGBA(c, *(rgba + [self._conn.SERVICE_OPTS]))
+            if (c+1 in abs_channels):
+                idx += 1
         return True
 
     def getProjections (self):
@@ -6371,19 +6408,19 @@ class _ImageWrapper (BlitzObjectWrapper):
 
 
     @assert_re()
-    def renderJpeg (self, z, t, compression=0.9):
+    def renderJpeg (self, z=None, t=None, compression=0.9):
         """
         Return the data from rendering image, compressed (and projected).
         Projection (or not) is specified by calling L{setProjection} before renderJpeg.
         
-        @param z:               The Z index. Ignored if projecting image. 
-        @param t:               The T index. 
+        @param z:               The Z index. Ignored if projecting image. If None, use defaultZ
+        @param t:               The T index. If None, use defaultT
         @param compression:     Compression level for jpeg
         @type compression:      Float
         """
         
-        self._pd.z = long(z)
-        self._pd.t = long(t)
+        self._pd.z = z is not None and long(z) or self._re.getDefaultZ()
+        self._pd.t = t is not None and long(t) or self._re.getDefaultT()
         try:
             if compression is not None:
                 try:
@@ -6596,7 +6633,12 @@ class _ImageWrapper (BlitzObjectWrapper):
         for chunk in ofw.getFileInChunks():
             outfile.write(chunk)
         outfile.close()
-        self._conn.deleteObjects('/OriginalFile', todel) # No error handling?
+        handle = self._conn.deleteObjects('/OriginalFile', todel)
+        try:
+            self._conn._waitOnCmd(handle)
+        finally:
+            handle.close()
+
         return os.path.splitext(f.name.val)[-1], f.mimetype.val
         
     def renderImage (self, z, t, compression=0.9):
@@ -7016,7 +7058,10 @@ class _ImageWrapper (BlitzObjectWrapper):
 
     def _deleteSettings(self):
         handle = self._conn.deleteObjects("/Image/Pixels/RenderingDef", [self.getId()])
-        self._conn._waitOnCmd(handle)
+        try:
+            self._conn._waitOnCmd(handle)
+        finally:
+            handle.close()
 
     def _collectRenderOptions (self):
         """
